fix decode-type mismatches surfaced by hasql-interpolate strictness
apis.issues.seq_num is INT; cast to bigint in the issue-list SELECT so
the generic DecodeRow doesn't blow up on int4-vs-int8.

projects.query_library.query_type is an enum; cast to text in the three
UNION branches of queryLibHistoryForUser so the generic decoder reads it
as Text instead of a custom OID.

Add migration 0096 to convert the four remaining VARCHAR(n) columns
(users.first_name/last_name, projects.teams.name/handle) to TEXT.
VARCHAR's only practical difference is the catalog OID, and that OID
mismatch was crashing every team-list and member-list lookup under the
new hasql decoders.

+++ src/Models/Apis/Issues.hs
@@ -448,7 +448,7 @@ selectIssues pid _typeM isAcknowledged isArchived limit offset timeRangeM sortM 
         i.affected_requests::bigint, i.affected_clients::bigint, NULL::double precision,
         i.recommended_action, i.migration_complexity, i.issue_data, i.request_payloads, i.response_payloads,
         NULL::timestamp with time zone, NULL::bigint,
-        i.target_hash, NULL::text, i.seq_num, i.parent_hash, i.is_framework,
+        i.target_hash, NULL::text, i.seq_num::bigint, i.parent_hash, i.is_framework,
         CASE
           WHEN i.issue_type = 'runtime_exception' THEN COALESCE(err_ev.cnt, 0)
           WHEN i.issue_type IN ('log_pattern', 'log_pattern_rate_change') THEN COALESCE(lp_ev.cnt, 0)
